Fix item selection in graphics view
Drop GraphElement::m_selected as it is not needed, QAbstractGraphicsItem
covers that. It is not necessary for the graph model to know about selected
items, only for the view.
Also remove redundant logic about selecting multiple elements in the
view, QGraphicsView is already capable of this.

This commit fixes item (un)selection and multiple selection logic