Fix some issues with threading.
There were potential race conditions where the next graph
could be processed before we could read out the results
from the previous one.
That is probably more of a theoretical issue but it
seemed best to fix it anyway.
A more serious issue is that graphviz obviously is not
thread-safe but actually crashes frequently, so wrap
the problematic functions in mutexes.

Signed-off-by: Reimar Döffinger <Reimar.Doeffinger@gmx.de>