Bump GraphViz requirements to 2.30
The switch from libgraph to libcgraph happened in version 2.30,
released on January 2013. Distributions released in 2014 or later
which provides GraphViz packages ship versions newer than 2.30.
Newer builds are available also from the GraphViz website.
No need to support such older versions.

Remove the patch (which should have been committed in a separate
branch, but it's not important anymore) which supported the old
version based on libgraph.