Revert "Revert "No keyword in part factory create()""
This reverts commit caa232d767a16f6033193d16d4b0e715fd06708b.

Plain KPart users like Konqueror or KDevelop do not use a keyword
when asking the factory for a plugin. Given there is also only one
plugin and no other yet planned, a keyword is not really needed,
and any users of the part need to port their related code to Qt5/KF5
anyway, so removing the keyword will not break anything.