Switch KgvPageLayout to QPageSize for page size handling
The current code in KgvPageLayout has a list of page sizes, with own
translations and measures (with even possible mistakes in sizes).

Switch KgvPageLayout to use QPageSize exclusively, leaving only the
KgvFormat <-> QPageSize::PageSizeId mappings, and a couple of custom
descriptions.  There are no resulting behaviour changes, with only the
page sizes that are different (since they come from Qt).