Remove broken argument SERVICE_TYPES kpart.desktop from desktop2json call
That was missing the needed definition for KParts/ReadOnlyPart,
so desktop2json was falling back to gueesing the file by the name.

That was failing for KF <5.75, as the type definition file names were
not matching the type name. Fixed in newer versions. For those older ones,
the warning output can just be ignored as it had been by everyone so far.

GIT_SILENT